# swagger.yaml
openapi: 3.0.0
servers:
  - description: QuitSmoke API
    url: http://localhost:5000
info:
  version: "1.0.0"
  title: QuitSmoke API by Hoàng Đức
  description: API for managing user authentication, blogs, profiles, surveys, memberships, payments, chat, and quit plans, developed by Hoàng Đức.
  contact:
    name: Hoàng Đức
    email: hoangduc@example.com
security:
  - BearerAuth: []
paths:
...
  /plans/quitplans:
    get:
      tags:
        - QuitPlan
      summary: Get quit plans
      description: Retrieves a list of quit plans filtered by user ID, coach ID, or status with pagination.
      operationId: getQuitPlans
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: query
          description: Filter by user ID
          schema:
            type: string
            example: '12345'
        - name: coachId
          in: query
          description: Filter by coach ID
          schema:
            type: string
            example: '67890'
        - name: status
          in: query
          description: Filter by quit plan status
          schema:
            type: string
            enum:
              - template
              - pending
              - ongoing
              - completed
              - failed
            example: 'ongoing'
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            example: 1
        - name: limit
          in: query
          description: Number of quit plans per page
          schema:
            type: integer
            minimum: 1
            example: 10
      responses:
        '200':
          description: List of quit plans with pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  quitPlans:
                    type: array
                    items:
                      $ref: '#/components/schemas/QuitPlan'
                  total:
                    type: integer
                    example: 50
                  page:
                    type: integer
                    example: 1
                  limit:
                    type: integer
                    example: 10
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid query parameters (e.g., invalid userId, coachId, or status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - QuitPlan
      summary: Create a new quit plan template
      description: Creates a new quit plan template (requires coach role).
      operationId: createQuitPlan
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuitPlanCreateRequest'
        required: true
      responses:
        '201':
          description: Quit plan template created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuitPlan'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - User is not a coach
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid request data (e.g., missing title or reason)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /plans/quitplans/{id}:
    get:
      tags:
        - QuitPlan
      summary: Get quit plan by ID
      description: Retrieves a quit plan and its associated badges by ID.
      operationId: getQuitPlanById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the quit plan
          schema:
            type: string
            example: '12345'
      responses:
        '200':
          description: Quit plan and associated badges
          content:
            application/json:
              schema:
                type: object
                properties:
                  quitPlan:
                    $ref: '#/components/schemas/QuitPlan'
                  badges:
                    type: array
                    items:
                      $ref: '#/components/schemas/Badge'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quit plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - QuitPlan
      summary: Update a quit plan
      description: Updates an existing quit plan (requires coach or user ownership).
      operationId: updateQuitPlan
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the quit plan
          schema:
            type: string
            example: '12345'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuitPlanUpdateRequest'
        required: true
      responses:
        '200':
          description: Quit plan updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuitPlan'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - User does not own the plan or is not the coach
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quit plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - QuitPlan
      summary: Delete a quit plan
      description: Deletes a quit plan and its associated stages and badges (requires coach ownership).
      operationId: deleteQuitPlan
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the quit plan
          schema:
            type: string
            example: '12345'
      responses:
        '200':
          description: Quit plan deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - User is not the coach of the plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quit plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /plans/quitplans/{quitPlanId}/stages:
    get:
      tags:
        - QuitPlanStage
      summary: Get quit plan stages
      description: Retrieves all stages for a specific quit plan with pagination.
      operationId: getQuitPlanStages
      security:
        - BearerAuth: []
      parameters:
        - name: quitPlanId
          in: path
          required: true
          description: ID of the quit plan
          schema:
            type: string
            example: '12345'
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            example: 1
        - name: limit
          in: query
          description: Number of stages per page
          schema:
            type: integer
            minimum: 1
            example: 10
      responses:
        '200':
          description: List of quit plan stages with pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  stages:
                    type: array
                    items:
                      $ref: '#/components/schemas/QuitPlanStage'
                  total:
                    type: integer
                    example: 20
                  page:
                    type: integer
                    example: 1
                  limit:
                    type: integer
                    example: 10
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid query parameters (e.g., invalid quitPlanId)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quit plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - QuitPlanStage
      summary: Create a new quit plan stage
      description: Creates a new stage for a quit plan (requires coach role).
      operationId: createQuitPlanStage
      security:
        - BearerAuth: []
      parameters:
        - name: quitPlanId
          in: path
          required: true
          description: ID of the quit plan
          schema:
            type: string
            example: '12345'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuitPlanStageCreateRequest'
        required: true
      responses:
        '201':
          description: Quit plan stage created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuitPlanStage'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - User is not a coach
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid request data (e.g., missing stage_name or order_index)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quit plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /plans/quitplan-stages/{id}:
    put:
      tags:
        - QuitPlanStage
      summary: Update a quit plan stage
      description: Updates an existing quit plan stage (requires coach role).
      operationId: updateQuitPlanStage
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the quit plan stage
          schema:
            type: string
            example: 'stage-12345'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuitPlanStageUpdateRequest'
        required: true
      responses:
        '200':
          description: Quit plan stage updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuitPlanStage'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - User is not a coach
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quit plan stage not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - QuitPlanStage
      summary: Delete a quit plan stage
      description: Deletes a quit plan stage (requires coach ownership).
      operationId: deleteQuitPlanStage
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the quit plan stage
          schema:
            type: string
            example: 'stage-12345'
      responses:
        '200':
          description: Quit plan stage deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - User is not the coach of the plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quit plan stage not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /plans/quitplans/{quitPlanId}/badges:
    get:
      tags:
        - Badge
      summary: Get quit plan badges
      description: Retrieves all badges for a specific quit plan with pagination.
      operationId: getQuitPlanBadges
      security:
        - BearerAuth: []
      parameters:
        - name: quitPlanId
          in: path
          required: true
          description: ID of the quit plan
          schema:
            type: string
            example: '12345'
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            example: 1
        - name: limit
          in: query
          description: Number of badges per page
          schema:
            type: integer
            minimum: 1
            example: 10
      responses:
        '200':
          description: List of badges with pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  badges:
                    type: array
                    items:
                      $ref: '#/components/schemas/Badge'
                  total:
                    type: integer
                    example: 20
                  page:
                    type: integer
                    example: 1
                  limit:
                    type: integer
                    example: 10
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quit plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Badge
      summary: Award a badge to a quit plan
      description: Awards a badge to a quit plan user (requires coach role).
      operationId: awardBadge
      security:
        - BearerAuth: []
      parameters:
        - name: quitPlanId
          in: path
          required: true
          description: ID of the quit plan
          schema:
            type: string
            example: '12345'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BadgeCreateRequest'
        required: true
      responses:
        '201':
          description: Badge awarded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Badge'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - User is not a coach
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quit plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid request data (e.g., missing name or description)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /plans/quitplans/select:
    post:
      tags:
        - QuitPlan
      summary: Select a template quit plan
      description: Allows a user to select and activate a template quit plan created by a coach.
      operationId: selectQuitPlan
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - quitPlanId
              properties:
                quitPlanId:
                  type: string
                  example: 'template-plan-123'
                  description: ID of the template plan to select
        required: true
      responses:
        '201':
          description: Quit plan selected and created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuitPlan'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid request (e.g., user already has an ongoing plan, invalid template, or selecting own template)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /plans/quitplans/current:
    get:
      tags:
        - QuitPlan
      summary: Get current quit plan
      description: Retrieves the current ongoing quit plan for the authenticated user, including stages and progress.
      operationId: getUserCurrentPlan
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current quit plan with stages and progress, or null if no ongoing plan exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    $ref: '#/components/schemas/QuitPlan'
                    nullable: true
                  stages:
                    type: array
                    items:
                      $ref: '#/components/schemas/QuitPlanStage'
                  progress:
                    type: array
                    items:
                      $ref: '#/components/schemas/QuitProgress'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /plans/quitplan-stages/{id}/complete:
    put:
      tags:
        - QuitPlanStage
      summary: Complete a quit plan stage (Manual)
      description: Manually marks a quit plan stage as completed. Note - Stages are normally auto-completed when users log daily progress for all required days. If all stages are completed, the plan is marked as completed and a badge is awarded.
      operationId: completeStage
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the quit plan stage
          schema:
            type: string
            example: 'stage-12345'
      responses:
        '200':
          description: Stage completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Stage completed successfully'
                  planCompleted:
                    type: boolean
                    example: false
                    description: Indicates if the entire plan was completed
                  completedStages:
                    type: integer
                    example: 3
                    description: Number of stages completed
                  totalStages:
                    type: integer
                    example: 5
                    description: Total number of stages in the plan
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Stage or quit plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid request (e.g., stage not part of user's ongoing plan)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /plans/quitplans/{planId}/complete:
    put:
      tags:
        - QuitPlan
      summary: Complete a quit plan
      description: Manually marks a quit plan as completed and awards a badge to the user.
      operationId: completePlan
      security:
        - BearerAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          description: ID of the quit plan to complete
          schema:
            type: string
            example: '12345'
      responses:
        '200':
          description: Plan completed successfully and badge awarded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Plan completed successfully! Badge awarded.'
                  data:
                    type: object
                    properties:
                      plan:
                        $ref: '#/components/schemas/QuitPlan'
                      badge:
                        $ref: '#/components/schemas/Badge'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quit plan not found or not ongoing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid request (e.g., plan not owned by user or not in ongoing status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /plans/quitplans/{id}/fail:
    put:
      tags:
        - QuitPlan
      summary: Mark quit plan as failed
      description: Allows a user to mark their ongoing quit plan as failed.
      operationId: failQuitPlan
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the quit plan
          schema:
            type: string
            example: '12345'
      responses:
        '200':
          description: Quit plan marked as failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuitPlan'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quit plan not found or not ongoing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    QuitPlan:
      type: object
      properties:
        _id:
          type: string
          example: '12345'
        coachId:
          type: object
          properties:
            _id:
              type: string
              example: '67890'
            name:
              type: string
              example: 'Coach John'
            email:
              type: string
              example: 'john@example.com'
        userId:
          type: object
          nullable: true
          properties:
            _id:
              type: string
              example: '54321'
            name:
              type: string
              example: 'Jane Doe'
            email:
              type: string
              example: 'jane@example.com'
        title:
          type: string
          example: 'Smoking Cessation Plan'
        reason:
          type: string
          example: 'Improve health and save money'
        startDate:
          type: string
          format: date-time
          example: '2025-06-02T13:46:00.000Z'
        endDate:
          type: string
          format: date-time
          example: '2025-07-02T13:46:00.000Z'
        status:
          type: string
          enum: ['template', 'pending', 'ongoing', 'completed', 'failed']
          example: 'template'
        createdAt:
          type: string
          format: date-time
          example: '2025-06-02T13:46:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2025-06-02T13:46:00.000Z'
      required:
        - coachId
        - title
        - reason
        - status
    QuitPlanCreateRequest:
      type: object
      properties:
        title:
          type: string
          example: 'Smoking Cessation Plan'
        reason:
          type: string
          example: 'Improve health and save money'
        startDate:
          type: string
          format: date-time
          example: '2025-06-02T13:46:00.000Z'
        endDate:
          type: string
          format: date-time
          example: '2025-07-02T13:46:00.000Z'
      required:
        - title
        - reason
...
    Badge:
      type: object
      properties:
        _id:
          type: string
          example: 'badge-12345'
        quitPlanId:
          type: string
          example: '12345'
        userId:
          type: object
          properties:
            _id:
              type: string
              example: '54321'
            name:
              type: string
              example: 'Jane Doe'
            email:
              type: string
              example: 'jane@example.com'
        name:
          type: string
          example: 'Stage Completed'
        description:
          type: string
          example: 'Completed stage: Week 1'
        awardedAt:
          type: string
          format: date-time
          example: '2025-06-02T13:46:00.000Z'
      required:
        - quitPlanId
        - userId
        - name
        - awardedAt
...
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          example: 'success'
        message:
          type: string
          example: 'Operation completed successfully'
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: 'error'
        message:
          type: string
          example: 'Invalid request data'
...